test: split scheduler and review comment tests
